Basic input/output of football graph
