--- .idea/.name
@@ -1,1 +1,0 @@
-girvan_newman_kaspereroy
